+++ parser_library/include/processing_format.h
@@ -48,8 +48,9 @@ enum class operand_occurrence : uint8_t
 // structure respresenting in which fashion should be statement processed
 struct processing_format
 {
-    constexpr processing_format(
-        processing_kind kind, processing_form form, operand_occurrence occurrence = operand_occurrence::PRESENT)
+    constexpr processing_format(processing_kind kind,
+        processing_form form,
+        operand_occurrence occurrence = operand_occurrence::PRESENT) noexcept
         : kind(kind)
         , form(form)
         , occurrence(occurrence)